% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ParallelpamSetDebug}
\alias{ParallelpamSetDebug}
\title{ParallelpamSetDebug}
\usage{
ParallelpamSetDebug(deb = TRUE, debjmat = FALSE)
}
\arguments{
\item{deb}{boolean, TRUE to generate debug messages for the PAM algorithm and silhouette calculation and FALSE to turn them off. Default: true.}

\item{debjmat}{boolean, TRUE to generate debug messages for the jmatrix part inside this package and FALSE to turn them off. Default: false}
}
\value{
No return value, called for side effects (internal boolean flag changed)
}
\description{
Sets debugging in parallelpam package to ON (with TRUE) or OFF (with FALSE) for both parts of it.\cr
On package load the default status is OFF.\cr
Setting debugging of any part to ON shows a message. Setting to OFF does not show anything (since debugging is OFF...)
}
\examples{
ParallelpamSetDebug(TRUE,debjmat=TRUE)
ParallelpamSetDebug(TRUE,debjmat=FALSE)
}
